requirements.txt
flask==2.3.3
anthropic==0.58.2
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.11.7
pytest==8.4.1
selenium==4.34.2
webdriver-manager==4.0.2